(QA-3437) Move perf setup code into helper
