+++ setup/install_gatling/00_pre_install/05_initialize_helpers.rb
@@ -1,0 +1,3 @@
+require File.expand_path('../../../../setup/helpers/perf_helper', __FILE__)
+Beaker::TestCase.class_eval { include PerfHelper }
+perf_init
