+++ setup/install_gatling/00_pre_install/10_etc_hosts.rb
@@ -1,7 +1,5 @@
 test_name 'set loadbalancer as puppet in /etc/hosts' do
   step 'modify /etc/hosts' do
-    puppet_ip = any_hosts_as?(:loadbalancer) ? loadbalancer.ip : master.ip
-    on hosts, %Q{echo "#{puppet_ip} puppet" >> /etc/hosts}
+    set_etc_hosts
   end
-
 end
