+++ setup/install_gatling/40_post_install/60_restart_server.rb
@@ -4,4 +4,4 @@ service_name = get_puppet_server_service_name_from_env()
 
 on(master, "service #{service_name} reload")
 
-Beaker::Log.notify("Finished restarting service #{service_name}")
+logger.notify("Finished restarting service #{service_name}")
